fix: корректировка логики агрегатора, double вместо int для weight

+++ telemetry/analyzer/src/main/java/ru/practicum/analyzer/repository/UserActionRepository.java
@@ -18,15 +18,13 @@ public interface UserActionRepository extends JpaRepository<UserAction, Long> {
     @Query("SELECT ua.eventId FROM UserAction ua WHERE ua.userId = :userId")
     List<Long> findEventIdsByUserId(@Param("userId") Long userId);
 
-    // Получение последних действий пользователя
     @Query("SELECT ua FROM UserAction ua WHERE ua.userId = :userId ORDER BY ua.timestamp DESC")
     List<UserAction> findLastUserActions(@Param("userId") Long userId, Pageable pageable);
 
-    // Сумма максимальных весов для каждого пользователя (исправлено)
+    // Возвращаем Double
     @Query(value = "SELECT COALESCE(SUM(max_weight), 0) FROM (SELECT MAX(weight) as max_weight FROM user_actions WHERE event_id = :eventId GROUP BY user_id) sub", nativeQuery = true)
-    Long sumMaxWeightsByEventId(@Param("eventId") Long eventId);
+    Double sumMaxWeightsByEventId(@Param("eventId") Long eventId);
 
-    // Получение всех действий по мероприятию с их весами
     @Query("SELECT ua FROM UserAction ua WHERE ua.eventId = :eventId")
     List<UserAction> findByEventId(@Param("eventId") Long eventId);
 }
